requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Django==4.2.3
django-cors-headers==4.2.0
django-paypal==2.0
djangorestframework==3.14.0
idna==3.4
pycparser==2.21
pyOpenSSL==23.2.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
